requirements.txt
azure-functions
pytz
requests